+++ pycep/typing.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from typing_extensions import Literal, TypeAlias, TypedDict
+from typing_extensions import Literal, NotRequired, TypeAlias, TypedDict
 
 # dict[str, Any] -> dict[str, PossibleValue] not supported https://github.com/python/mypy/issues/731
 PossibleValue: TypeAlias = "bool | int | str | list[bool | int | str] | dict[str, Any]"
@@ -54,8 +54,8 @@ class Globals(TypedDict):
 
 class ScopeAttributes(TypedDict):
     value: Literal["resourceGroup", "subscription", "managementGroup", "tenant"]
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class _Scope(TypedDict):
@@ -71,8 +71,8 @@ class ParameterAttributes(TypedDict):
     decorators: list[Decorator]
     type: str
     default: PossibleNoneValue
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class _Parameters(TypedDict):
@@ -87,8 +87,8 @@ class ParamResponse(TypedDict):
 class VariableAttributes(TypedDict):
     decorators: list[Decorator]
     value: PossibleValue
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class _Variables(TypedDict):
@@ -104,8 +104,8 @@ class OutputAttributes(TypedDict):
     decorators: list[Decorator]
     type: str
     value: PossibleValue
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class _Outputs(TypedDict):
@@ -123,8 +123,8 @@ class ResourceAttributes(TypedDict):
     api_version: str
     existing: bool
     config: dict[str, Any]
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class Resource(TypedDict):
@@ -141,8 +141,8 @@ class ModuleAttributes(TypedDict):
     type: Literal["local", "bicep_registry", "bicep_registry_alias", "template_spec", "template_spec_alias"]
     detail: ModuleDetail
     config: dict[str, Any]
-    __start_line__: int | None  # NotRequired[int]
-    __end_line__: int | None  # NotRequired[int]
+    __start_line__: NotRequired[int]
+    __end_line__: NotRequired[int]
 
 
 class _Modules(TypedDict):
@@ -413,22 +413,22 @@ class Empty(TypedDict):
     parameters: _EmptyParameters
 
 
-class First(TypedDict, total=False):
+class First(TypedDict):
     type: Literal["first"]
     parameters: _LengthParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
-class Intersection(TypedDict, total=False):
+class Intersection(TypedDict):
     type: Literal["intersection"]
     parameters: _UnionParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
-class Last(TypedDict, total=False):
+class Last(TypedDict):
     type: Literal["last"]
     parameters: _LengthParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 class _LengthParameters(TypedDict):
@@ -470,16 +470,16 @@ class Take(TypedDict):
     parameters: _TakeParameters
 
 
-class _UnionParameters(TypedDict, total=False):
+class _UnionParameters(TypedDict):
     arg_1: str
     arg_2: str
-    arg_3: str  # NotRequired[str]  # and many more possible
+    arg_3: NotRequired[str]  # and many more possible
 
 
-class UnionFunc(TypedDict, total=False):
+class UnionFunc(TypedDict):
     type: Literal["union"]
     parameters: _UnionParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 ####################
@@ -516,14 +516,14 @@ class UtcNow(TypedDict):
 ####################
 
 
-class Deployment(TypedDict, total=False):
+class Deployment(TypedDict):
     type: Literal["deployment"]
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
-class Environment(TypedDict, total=False):
+class Environment(TypedDict):
     type: Literal["environment"]
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 ####################
@@ -542,7 +542,7 @@ class LoadFileAsBase64(TypedDict):
     parameters: _LoadFileAsBase64Parameters
 
 
-class _LoadTextContentParameters(TypedDict, total=False):
+class _LoadTextContentParameters(TypedDict):
     file_path: PossibleValue
     encoding: PossibleNoneValue
 
@@ -616,10 +616,10 @@ class _ListKeysParameters(TypedDict):
     api_version: PossibleValue
 
 
-class ListKeys(TypedDict, total=False):
+class ListKeys(TypedDict):
     type: Literal["list_keys"]
     parameters: _ListKeysParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 class _PickZonesParameters(TypedDict):
@@ -641,10 +641,10 @@ class _ReferenceParameters(TypedDict):
     full: PossibleNoneValue
 
 
-class Reference(TypedDict, total=False):
+class Reference(TypedDict):
     type: Literal["reference"]
     parameters: _ReferenceParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 class _ResourceIdParameters(TypedDict):
@@ -694,10 +694,10 @@ class _ManagementGroupParameters(TypedDict):
     identifier: PossibleNoneValue
 
 
-class ManagementGroup(TypedDict, total=False):
+class ManagementGroup(TypedDict):
     type: Literal["management_group"]
     parameters: _ManagementGroupParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 class _ResourceGroupParameters(TypedDict):
@@ -705,25 +705,25 @@ class _ResourceGroupParameters(TypedDict):
     subscription_id: PossibleNoneValue
 
 
-class ResourceGroup(TypedDict, total=False):
+class ResourceGroup(TypedDict):
     type: Literal["resource_group"]
     parameters: _ResourceGroupParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 class _SubscriptionParameters(TypedDict):
     subscription_id: PossibleNoneValue
 
 
-class Subscription(TypedDict, total=False):
+class Subscription(TypedDict):
     type: Literal["subscription"]
     parameters: _SubscriptionParameters
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
-class Tenant(TypedDict, total=False):
+class Tenant(TypedDict):
     type: Literal["tenant"]
-    property_name: str  # NotRequired[str]
+    property_name: NotRequired[str]
 
 
 ####################
@@ -779,10 +779,10 @@ class EndsWith(TypedDict):
     parameters: _StartsWithParameters
 
 
-class _FormatParameters(TypedDict, total=False):
+class _FormatParameters(TypedDict):
     format_string: PossibleValue
     arg_1: PossibleValue
-    arg_2: PossibleValue  # NotRequired[PossibleValue]  # and many more possible
+    arg_2: NotRequired[PossibleValue]  # and many more possible
 
 
 class Format(TypedDict):
@@ -790,9 +790,9 @@ class Format(TypedDict):
     parameters: _FormatParameters
 
 
-class _GuidParameters(TypedDict, total=False):
+class _GuidParameters(TypedDict):
     base_string: str
-    extra_string_1: str  # NotRequired[str]  # and many more possible
+    extra_string_1: NotRequired[str]  # and many more possible
 
 
 class Guid(TypedDict):
@@ -846,10 +846,10 @@ class _SplitParameters(TypedDict):
     delimiter: PossibleValue
 
 
-class Split(TypedDict, total=False):
+class Split(TypedDict):
     type: Literal["split"]
     parameters: _SplitParameters
-    index: int  # NotRequired[int]
+    index: NotRequired[int]
 
 
 class _StartsWithParameters(TypedDict):
@@ -1008,10 +1008,10 @@ class NotEqualsCaseInsensitive(TypedDict):
 ####################
 
 
-class _AndOperands(TypedDict, total=False):
+class _AndOperands(TypedDict):
     operand_1: PossibleValue
     operand_2: PossibleValue
-    operand_3: PossibleValue  # NotRequired[PossibleValue]  # and many more possible
+    operand_3: NotRequired[PossibleValue]  # and many more possible
 
 
 class And(TypedDict):
@@ -1102,10 +1102,10 @@ class Substract(TypedDict):
 ####################
 
 
-class BicepJson(TypedDict, total=False):
+class BicepJson(TypedDict):
     globals: GlobalsAttributes
-    parameters: dict[str, ParameterAttributes]  # NotRequired[dict[str, ParameterAttributes]]
-    variables: dict[str, VariableAttributes]  # NotRequired[dict[str, VariableAttributes]]
-    resources: dict[str, ResourceAttributes]  # NotRequired[dict[str, ResourceAttributes]]
-    modules: dict[str, ModuleAttributes]  # NotRequired[dict[str, ModuleAttributes]]
-    outputs: dict[str, OutputAttributes]  # NotRequired[dict[str, OutputAttributes]]
+    parameters: NotRequired[dict[str, ParameterAttributes]]
+    variables: NotRequired[dict[str, VariableAttributes]]
+    resources: NotRequired[dict[str, ResourceAttributes]]
+    modules: NotRequired[dict[str, ModuleAttributes]]
+    outputs: NotRequired[dict[str, OutputAttributes]]